fix datanumber converter, streetadress compundtype mapping, and formula

+++ src/Nox.Types.EntityFramework/Types/DatabaseNumber/DatabaseNumberConverter.cs
@@ -2,7 +2,7 @@
 
 namespace Nox.Types.EntityFramework.Types;
 
-public class DatabaseNumberConverter : ValueConverter<DatabaseNumber, uint>
+public class DatabaseNumberConverter : ValueConverter<DatabaseNumber, ulong>
 {
     public DatabaseNumberConverter() : base(databaseNumber => databaseNumber.Value, currentNumber => DatabaseNumber.FromDatabase(currentNumber)) { }
 }
